requirements.txt
# =============================================================================
#  ʟᴇᴇᴄʜʙᴏᴛ - ᴀᴅᴠᴀɴᴄᴇᴅ ᴛᴇʟᴇɢʀᴀᴍ ғɪʟᴇ ᴛʀᴀɴsʟᴏᴀᴅᴇʀ
# =============================================================================
#  ᴄᴏᴘʏʀɪɢʜᴛ © 2024-2025 sʜɪɴᴇɪ ɴᴏᴜᴢᴇɴ
#  ɢɪᴛʜᴜʙ: https://ɢɪᴛʜᴜʙ.ᴄᴏᴍ/sʜɪɴᴇɪɪ86
#  ᴛᴇʟᴇɢʀᴀᴍ: https://ᴛ.ᴍᴇ/sʜɪɴᴇɪɪ86
# =============================================================================

# ᴄᴏʀᴇ ᴅᴇᴘᴇɴᴅᴇɴᴄɪᴇs
pyrogram>=2.0.106
tgcrypto>=1.2.5

# ᴅᴏᴡɴʟᴏᴀᴅᴇʀs
yt-dlp>=2024.12.23
aiohttp>=3.9.0
requests>=2.31.0

# ɢᴏᴏɢʟᴇ ᴅʀɪᴠᴇ
google-api-python-client>=2.115.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ᴠɪᴅᴇᴏ ᴘʀᴏᴄᴇssɪɴɢ
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# ɪᴍᴀɢᴇ ᴘʀᴏᴄᴇssɪɴɢ
Pillow>=10.2.0
opencv-python>=4.9.0.80

# ᴜᴛɪʟɪᴛɪᴇs
psutil>=5.9.8
GPUtil>=1.4.0
natsort>=8.4.0
pytz>=2024.1

# ᴘᴇʀғᴏʀᴍᴀɴᴄᴇ
uvloop>=0.19.0

# ᴏᴘᴛɪᴏɴᴀʟ ᴅᴇᴘᴇɴᴅᴇɴᴄɪᴇs
# ᴍᴇɢᴀᴛᴏᴏʟs (ɪɴsᴛᴀʟʟ ᴠɪᴀ ᴀᴘᴛ: ᴀᴘᴛ ɪɴsᴛᴀʟʟ ᴍᴇɢᴀᴛᴏᴏʟs)
# ᴀʀɪᴀ2 (ɪɴsᴛᴀʟʟ ᴠɪᴀ ᴀᴘᴛ: ᴀᴘᴛ ɪɴsᴛᴀʟʟ ᴀʀɪᴀ2)
# ғғᴍᴘᴇɢ (ɪɴsᴛᴀʟʟ ᴠɪᴀ ᴀᴘᴛ: ᴀᴘᴛ ɪɴsᴛᴀʟʟ ғғᴍᴘᴇɢ)
